fix: Use root in serveStatic so that the script can work from anywhere
Previously, if the node script is at ~/.npm/_npx/some/folders/index.js but you
run it from cwd eg. ~/Documents, the relativePathToScript will start with "../"
which hono serveStatic does not expect as part of the "path" param. This leads to
the file not being found.

This fixes the problem by moving the part with "../" to the "root" param, where
it gets resolved correctly.